Moved some logic from fragment to activity in order to decouple the fragment from the activity
